Remove wrapper. psutil was actually only used by system metrics module and gateway, both are already lazy loaded. Hence, just added some check and suggestion.
