+++ mlflow/system_metrics/metrics/cpu_monitor.py
@@ -1,7 +1,8 @@
 """Class for monitoring CPU stats."""
 
+import psutil
+
 from mlflow.system_metrics.metrics.base_metrics_monitor import BaseMetricsMonitor
-from mlflow.utils import process
 
 
 class CPUMonitor(BaseMetricsMonitor):
@@ -9,10 +10,10 @@ class CPUMonitor(BaseMetricsMonitor):
 
     def collect_metrics(self):
         # Get CPU metrics.
-        cpu_percent = process.cpu_percent()
+        cpu_percent = psutil.cpu_percent()
         self._metrics["cpu_utilization_percentage"].append(cpu_percent)
 
-        system_memory = process.virtual_memory()
+        system_memory = psutil.virtual_memory()
         self._metrics["system_memory_usage_megabytes"].append(system_memory.used / 1e6)
         self._metrics["system_memory_usage_percentage"].append(
             system_memory.used / system_memory.total * 100
