+++ mlflow/system_metrics/metrics/disk_monitor.py
@@ -2,8 +2,9 @@
 
 import os
 
+import psutil
+
 from mlflow.system_metrics.metrics.base_metrics_monitor import BaseMetricsMonitor
-from mlflow.utils import process
 
 
 class DiskMonitor(BaseMetricsMonitor):
@@ -11,7 +12,7 @@ class DiskMonitor(BaseMetricsMonitor):
 
     def collect_metrics(self):
         # Get disk usage metrics.
-        disk_usage = process.disk_usage(os.sep)
+        disk_usage = psutil.disk_usage(os.sep)
         self._metrics["disk_usage_percentage"].append(disk_usage.percent)
         self._metrics["disk_usage_megabytes"].append(disk_usage.used / 1e6)
         self._metrics["disk_available_megabytes"].append(disk_usage.free / 1e6)
