+++ mlflow/utils/process.py
@@ -156,66 +156,3 @@ def cache_return_value_per_process(fn):
         return new_value
 
     return wrapped_fn
-
-
-"""
-Thin wrappers for psutil functions.
-As of 2023 Nov psutil doesn't have binary for Arm64 Linux, requireing building from CPython source. To
-avoid headache of building with GCC, we make psutil optional and show user-friendly message when it's
-required but not installed.
-"""
-
-
-def _check_and_import_psutil():
-    try:
-        import psutil
-    except ImportError as e:
-        raise ImportError(
-            str(e) + "\n\nHINT: The `psutil` package is required to log system metrics in "
-            "MLflow. Run `pip install psutil --upgrade` to install the package. If you are"
-            " on an Arm64-based Linux system, you may need to have GCC installed to build "
-            "the wheel from source."
-        )
-    return psutil
-
-
-def kill_child_processes(parent_pid):
-    """
-    Gracefully terminate or kill child processes from a main process
-    """
-    psutil = _check_and_import_psutil()
-
-    parent = psutil.Process(parent_pid)
-    for child in parent.children(recursive=True):
-        try:
-            child.terminate()
-        except psutil.NoSuchProcess:
-            pass
-    _, still_alive = psutil.wait_procs(parent.children(), timeout=3)
-    for p in still_alive:
-        p.kill()
-
-
-def cpu_count():
-    psutil = _check_and_import_psutil()
-    return psutil.cpu_count()
-
-
-def cpu_percent():
-    psutil = _check_and_import_psutil()
-    return psutil.cpu_percent()
-
-
-def virtual_memory():
-    psutil = _check_and_import_psutil()
-    return psutil.virtual_memory()
-
-
-def disk_usage(path: str):
-    psutil = _check_and_import_psutil()
-    return psutil.disk_usage(path)
-
-
-def net_io_counters():
-    psutil = _check_and_import_psutil()
-    return psutil.net_io_counters()
