+++ pylintrc
@@ -81,7 +81,6 @@ enable=signature-differs,
        # Custom rules
        lazy-builtin-import,
        useless-assignment,
-       illegal-direct-import,
 
 [REPORTS]
