+++ tests/gateway/tools.py
@@ -20,7 +20,7 @@ from sentence_transformers import SentenceTransformer
 
 import mlflow
 from mlflow.gateway import app
-from mlflow.utils.process import kill_child_processes
+from mlflow.gateway.utils import kill_child_processes
 
 from tests.helper_functions import _get_mlflow_home, _start_scoring_proc, get_safe_port
